backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
chromadb==0.4.15
PyPDF2==3.0.1
sentence-transformers==3.3.1
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.3
twilio==9.0.4
